% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_functions.r
\name{load_cube_projection}
\alias{load_cube_projection}
\title{Create a proxy data cube for future climate,
which loads data from a given image collection according to a data cube view based
on a specific box coordinates or using a set of observations}
\usage{
load_cube_projection(
  stac_path = "http://io.biodiversite-quebec.ca/stac/",
  limit = 5000,
  collections = c("chelsa-clim-proj"),
  use.obs = T,
  obs = NULL,
  lon = "lon",
  lat = "lat",
  buffer.box = 0,
  bbox = NULL,
  layers = NULL,
  variable = NULL,
  srs.cube = "EPSG:32198",
  time.span = "2041-2070",
  rcp = "ssp585",
  left = -2009488,
  right = 1401061,
  bottom = -715776,
  top = 2597757,
  spatial.res = 2000,
  temporal.res = "P1Y",
  aggregation = "mean",
  resampling = "near"
)
}
\arguments{
\item{stac_path, }{a character, base url of a STAC web service.}

\item{limit, }{an integer defining the maximum number of results to return.}

\item{collections, }{a character vector of collection IDs to include
subsetLayers, a vector, containing the name of layers to select. If NULL, all layers in dir.pred selected by default.}

\item{use.obs, }{a boolean. If TRUE, the provided observations will be sued as a basis for calculating the extent and bbox.}

\item{obs, }{a data.frame containg the observations (used if use.obs is T)}

\item{lon, }{a string, column from obs containing longitude}

\item{lat, }{a string, column from obs containing latitude}

\item{buffer.box, }{an integer, buffer to apply around the obs to calculate extent and bbox}

\item{bbox, }{a numeric vector of size 4 or 6. Coordinates of the bounding box (if use.obs is FALSE). Details in rstac::stac_search documentation.}

\item{layers, }{a string vector, names of bands to be used,. By default (NULL), all bands with "eo:bands" attributes will be used.}

\item{srs.cube, }{string, target spatial reference system. Can be a proj4 definition, WKT, or in the form "EPSG:XXXX".}

\item{time.span, }{a string, time interval of the projection model.}

\item{rcp, }{a string, climatic scenario}

\item{left, }{a float. Left coordinate of the extent. Used if use.obs = F}

\item{right, }{a float. Right coordinate of the extent. Used if use.obs = F}

\item{bottom, }{a float. Bottom coordinate of the extent. Used if use.obs = F}

\item{top, }{a float. Top coordinate of the extent. Used if use.obs = F}

\item{spatial.res, }{a float, size of pixels in longitude and latitude directions, in the unit of srs.cube spatial reference system.}

\item{temporal.res, }{size of pixels in time-direction, expressed as ISO8601 period string (only 1 number and unit is allowed) such as "P16D"}

\item{aggregation, }{a character, aggregation method as string, defining how to deal with pixels containing data from multiple images, can be "min", "max", "mean", "median", or "first"}

\item{resampling, }{a character, resampling method used in gdalwarp when images are read, can be "near", "bilinear", "bicubic" or others as supported by gdalwarp (see https://gdal.org/programs/gdalwarp.html)}

\item{srs.obs, }{string, observations spatial reference system. Can be a proj4 definition, WKT, or in the form "EPSG:XXXX".}
}
\value{
a raster stack of variables not intercorrelated

a proxy raster data cube
}
\description{
Create a proxy data cube for future climate,
which loads data from a given image collection according to a data cube view based
on a specific box coordinates or using a set of observations
}
