% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_functions.r
\name{load_prop_values}
\alias{load_prop_values}
\title{Create a proxy data cube for current climate,
which loads data from a given image collection according to a data cube view based
on a specific box coordinates or using a set of observations}
\usage{
load_prop_values(
  stac_path = "https://io.biodiversite-quebec.ca/stac/",
  collections = c("esacci-lc"),
  use.obs = T,
  obs = obs.coords.proj,
  buffer.box = 0,
  lat = "lat",
  lon = "lon",
  srs.cube = srs.cube,
  t0 = "2000-01-01",
  t1 = "2001-12-31",
  spatial.res = 250,
  prop = F,
  prop.res = 1000,
  select_values = NULL,
  layers = NULL,
  variable = NULL,
  temporal.res = "P1Y"
)
}
\arguments{
\item{stac_path, }{a character, base url of a STAC web service.}

\item{collections, }{a character vector of collection IDs to include
subsetLayers, a vector, containing the name of layers to select. If NULL, all layers in dir.pred selected by default.}

\item{use.obs, }{a boolean. If TRUE, the provided observations will be sued as a basis for calculating the extent and bbox.}

\item{obs, }{a data.frame containg the observations (used if use.obs is T)}

\item{buffer.box, }{an integer, buffer to apply around the obs to calculate extent and bbox}

\item{lat, }{a string, column from obs containing latitude}

\item{lon, }{a string, column from obs containing longitude}

\item{srs.cube, }{string, target spatial reference system. Can be a proj4 definition, WKT, or in the form "EPSG:XXXX".}

\item{t0, }{ISO8601 datetime string, start date.}

\item{t1, }{ISO8601 datetime string, end date.}

\item{spatial.res, }{a float, size of pixels in longitude and latitude directions, in the unit of srs.cube spatial reference system.}

\item{prop, }{a boolean. If TRUE, the proportion of each classes from selected_values is calculated. If FALSE, only the values in selected_values are retrieved.}

\item{prop.res, }{an integer, resolution to calculate the proportion of land cover classes (if prop is TRUE)}

\item{layers, }{a string vector, names of bands to be used,. By default (NULL), all bands with "eo:bands" attributes will be used.}

\item{temporal.res, }{size of pixels in time-direction, expressed as ISO8601 period string (only 1 number and unit is allowed) such as "P16D"}

\item{limit, }{an integer defining the maximum number of results to return.}

\item{srs.obs, }{string, observations spatial reference system. Can be a proj4 definition, WKT, or in the form "EPSG:XXXX".}

\item{bbox, }{a numeric vector of size 4 or 6. Coordinates of the bounding box (if use.obs is FALSE). Details in rstac::stac_search documentation.}

\item{left, }{a float. Left coordinate of the extent. Used if use.obs = F}

\item{right, }{a float. Right coordinate of the extent. Used if use.obs = F}

\item{top, }{a float. Top coordinate of the extent. Used if use.obs = F}

\item{bottom, }{a float. Bottom coordinate of the extent. Used if use.obs = F}

\item{selected_values, }{a vector, classes values to select.}
}
\value{
a raster stack of variables not intercorrelated

a proxy raster data cube
}
\description{
Create a proxy data cube for current climate,
which loads data from a given image collection according to a data cube view based
on a specific box coordinates or using a set of observations
}
